Extract isNodeFirstInLine to astUtil function
