+++ lib/rules/jsx-closing-tag-location.js
@@ -4,6 +4,8 @@
  */
 'use strict';
 
+const astUtil = require('../util/ast');
+
 // ------------------------------------------------------------------------------
 // Rule Definition
 // ------------------------------------------------------------------------------
@@ -18,31 +20,6 @@ module.exports = {
   },
 
   create: function(context) {
-    const sourceCode = context.getSourceCode();
-
-    /**
-     * Checks if the node is the first in its line, excluding whitespace.
-     * @param {ASTNode} node The node to check
-     * @return {Boolean} true if its the first node in its line
-     */
-    function isNodeFirstInLine(node) {
-      let token = node;
-      let lines;
-      do {
-        token = sourceCode.getTokenBefore(token);
-        lines = token.type === 'JSXText'
-          ? token.value.split('\n')
-          : null;
-      } while (
-        token.type === 'JSXText' &&
-        /^\s*$/.test(lines[lines.length - 1])
-      );
-
-      const startLine = node.loc.start.line;
-      const endLine = token ? token.loc.end.line : -1;
-      return startLine !== endLine;
-    }
-
     return {
       JSXClosingElement: function(node) {
         if (!node.parent) {
@@ -59,7 +36,7 @@ module.exports = {
         }
 
         let message;
-        if (!isNodeFirstInLine(node)) {
+        if (!astUtil.isNodeFirstInLine(context, node)) {
           message = 'Closing tag of a multiline JSX expression must be on its own line.';
         } else {
           message = 'Expected closing tag to match indentation of opening.';
@@ -71,7 +48,7 @@ module.exports = {
           message,
           fix: function(fixer) {
             const indent = Array(opening.loc.start.column + 1).join(' ');
-            if (isNodeFirstInLine(node)) {
+            if (astUtil.isNodeFirstInLine(context, node)) {
               return fixer.replaceTextRange(
                 [node.range[0] - node.loc.start.column, node.range[0]],
                 indent
