+++ lib/rules/jsx-indent.js
@@ -29,6 +29,8 @@
  */
 'use strict';
 
+const astUtil = require('../util/ast');
+
 // ------------------------------------------------------------------------------
 // Rule Definition
 // ------------------------------------------------------------------------------
@@ -153,29 +155,6 @@ module.exports = {
     }
 
     /**
-     * Checks node is the first in its own start line. By default it looks by start line.
-     * @param {ASTNode} node The node to check
-     * @return {Boolean} true if its the first in the its start line
-     */
-    function isNodeFirstInLine(node) {
-      let token = node;
-      let lines;
-      do {
-        token = sourceCode.getTokenBefore(token);
-        lines = token.type === 'JSXText'
-          ? token.value.split('\n')
-          : null;
-      } while (
-        token.type === 'JSXText' &&
-        /^\s*$/.test(lines[lines.length - 1])
-      );
-
-      const startLine = node.loc.start.line;
-      const endLine = token ? token.loc.end.line : -1;
-      return startLine !== endLine;
-    }
-
-    /**
      * Check if the node is the right member of a logical expression
      * @param {ASTNode} node The node to check
      * @return {Boolean} true if its the case, false if not
@@ -216,7 +195,7 @@ module.exports = {
       const isCorrectAlternateInCondExp = isAlternateInConditionalExp(node) && (nodeIndent - indent) === 0;
       if (
         nodeIndent !== indent &&
-        isNodeFirstInLine(node) &&
+        astUtil.isNodeFirstInLine(context, node) &&
         !isCorrectRightInLogicalExp &&
         !isCorrectAlternateInCondExp
       ) {
