+++ lib/util/ast.js
@@ -4,6 +4,30 @@
 'use strict';
 
 /**
+ * Find a return statment in the current node
+ *
+ * @param {ASTNode} ASTnode The AST node being checked
+ */
+function findReturnStatement(node) {
+  if (
+    (!node.value || !node.value.body || !node.value.body.body) &&
+    (!node.body || !node.body.body)
+  ) {
+    return false;
+  }
+
+  const bodyNodes = (node.value ? node.value.body.body : node.body.body);
+
+  let i = bodyNodes.length - 1;
+  for (; i >= 0; i--) {
+    if (bodyNodes[i].type === 'ReturnStatement') {
+      return bodyNodes[i];
+    }
+  }
+  return false;
+}
+
+/**
  * Get properties name
  * @param {Object} node - Property.
  * @returns {String} Property name.
@@ -36,31 +60,34 @@ function getComponentProperties(node) {
 }
 
 /**
- * Find a return statment in the current node
- *
- * @param {ASTNode} ASTnode The AST node being checked
- */
-function findReturnStatement(node) {
-  if (
-    (!node.value || !node.value.body || !node.value.body.body) &&
-    (!node.body || !node.body.body)
-  ) {
-    return false;
-  }
+     * Checks if the node is the first in its line, excluding whitespace.
+     * @param {Object} context The node to check
+     * @param {ASTNode} node The node to check
+     * @return {Boolean} true if its the first node in its line
+     */
+function isNodeFirstInLine(context, node) {
+  const sourceCode = context.getSourceCode();
+  let token = node;
+  let lines;
+  do {
+    token = sourceCode.getTokenBefore(token);
+    lines = token.type === 'JSXText'
+      ? token.value.split('\n')
+      : null;
+  } while (
+    token.type === 'JSXText' &&
+        /^\s*$/.test(lines[lines.length - 1])
+  );
 
-  const bodyNodes = (node.value ? node.value.body.body : node.body.body);
-
-  let i = bodyNodes.length - 1;
-  for (; i >= 0; i--) {
-    if (bodyNodes[i].type === 'ReturnStatement') {
-      return bodyNodes[i];
-    }
-  }
-  return false;
+  const startLine = node.loc.start.line;
+  const endLine = token ? token.loc.end.line : -1;
+  return startLine !== endLine;
 }
 
+
 module.exports = {
+  findReturnStatement: findReturnStatement,
   getPropertyName: getPropertyName,
   getComponentProperties: getComponentProperties,
-  findReturnStatement: findReturnStatement
+  isNodeFirstInLine: isNodeFirstInLine
 };
